requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
ecdsa==0.18.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-jose==3.3.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-core==4.4.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
